Revise extension memfrs
