+++ ext/memfrs/handles.h
@@ -17,6 +17,7 @@
  * License along with this library; if not, see <http://www.gnu.org/licenses/>.
  */
 
+
 typedef enum PARSING_HANDLE_TYPE{
     PARSING_HANDLE_TYPE_ALL,
     PARSING_HANDLE_TYPE_CR3,
@@ -28,3 +29,63 @@ typedef enum PARSING_HANDLE_TYPE{
 }PARSING_HANDLE_TYPE;
 
 extern PARSING_HANDLE_TYPE parsing_handle_type;
+
+
+// handles data sturcture
+typedef struct handles_node_st{
+    int handle_table_entry_index;
+    uint64_t handle_table_entry_address;
+    uint64_t grantedaccess;
+    char *type;
+    char *detail;
+}handles_node_st;
+
+typedef struct UT_handles{
+    uint64_t CR3;
+    uint64_t eprocess;
+    uint64_t pid;
+    char *imagename;
+    UT_array *handles_node;
+}handles_st;
+
+
+
+/*******************************************************************
+extern UT_array* memfrs_enum_proc_handles( int target_type, uint64_t target, uint64_t kpcr_ptr, CPUState *cpu )
+
+Eumerate the running process handles
+
+INPUT:     int target_type,          searching type of handles
+           uint64_t target,          searching target
+           uint64_t kpcr_ptr,        the address of _KPCR struct
+           CPUState *cpu,            the pointer to current cpu
+OUTPUT:    UT_array*                 return a UT_array with handles data
+*******************************************************************/
+extern UT_array* memfrs_enum_proc_handles( int target_type, uint64_t target, uint64_t kpcr_ptr, CPUState *cpu );
+
+
+/*******************************************************************
+extern UT_array* memfrs_enum_proc_handles_detail( int target_type, const char* target, uint64_t kpcr_ptr, CPUState *cpu )
+
+Eumerate the running process handles, expect for types and details
+
+INPUT:     int target_type,          searching type of handles
+           const char* target        searching target
+           uint64_t kpcr_ptr,        the address of _KPCR struct
+           CPUState *cpu,            the pointer to current cpu
+OUTPUT:    UT_array*                 return a UT_array with handles data
+*******************************************************************/
+extern UT_array* memfrs_enum_proc_handles_detail( int target_type, const char* target, uint64_t kpcr_ptr, CPUState *cpu );
+
+
+/*******************************************************************
+extern UT_array* memfrs_enum_handles_types( uint64_t kpcr_ptr, CPUState *cpu )
+
+Eumerate the handles types
+
+INPUT:     uint64_t kpcr_ptr,        the address of _KPCR struct
+           CPUState *cpu,            the pointer to current cpu
+OUTPUT:    UT_array*                 return a UT_array with handles types
+*******************************************************************/
+extern UT_array* memfrs_enum_handles_types( uint64_t kpcr_ptr, CPUState *cpu );
+
